Move webhook.controller into modules file
